airflowDags/requirements.txt
# Core Airflow dependencies
apache-airflow>=2.7.0
apache-airflow-providers-http>=4.5.0
apache-airflow-providers-postgres>=5.7.0
apache-airflow-providers-amazon>=8.4.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=14.0.0
boto3>=1.34.0

# Database
psycopg2-binary>=2.9.9
SQLAlchemy>=2.0.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
urllib3>=2.0.0

# Development tools
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0

# Note: Astro Runtime includes additional provider packages by default
# For more information, visit: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
